fix: conflict 해결 (#18)
