+++ src/main/java/com/nhnacademy/eggplantdeliveryschedule/repository/CustomDeliveryInfoRepository.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 public interface CustomDeliveryInfoRepository {
     List<String> retrieveStatusReadyTrackingNo();
+
     List<String> retrieveStatusDeliveringTrackingNo();
+
     List<String> retrieveStatusDeliveringFinalTrackingNo();
 }
